+++ 02-oop-inheritance/Actor.js
@@ -1,0 +1,19 @@
+export default class Actor {
+  constructor(name,age) {
+    this._name = name;
+    this._age = age;
+  }
+
+  get name() {
+    return this._name;
+  }
+  set name(name) {
+    this._name=name;
+  }
+  get age() {
+    return this._age;
+  }
+  set age(age) {
+    this._age=age;
+  }
+}
