+++ 02-oop-inheritance/Logger.js
@@ -1,0 +1,8 @@
+export default class Logger {
+  constructor() {
+
+  }
+  log(info) {
+    console.log(info);
+  }
+}
