+++ 02-oop-inheritance/index.html
@@ -6,6 +6,6 @@
   <title>Document</title>
 </head>
 <body>
-  <script src="movie.js"></script>
+  <script src="main.js"></script>
 </body>
 </html>
